Adding PHP 8 to test matrix, upgrading dev deps, fixing errors (#123)
* Adding PHP 8 to test matrix, upgrading dev deps, fixing errors

* disabling fast fail, adding merge-me job

* testing different parameter names in phpstan ignores

* more phpstan fixes

* more phpstan fixes, not there yet

* fixing what I can for php 8, rest seem harder to deal with

* upgrading phpstan, fixing php8 deprecation

* fixing case of DbTestCase file

* fixing php8 phpstan errors in Cache/Db

* fixing php8 phpstan errors in Connection.php

* fixing psalm, more php 8 fixes for Connection.php

* fixing php8 phpstan errors in Connection/Statement

* fixing more php 8 phpstan errors

* a few more php8 phpstan fixes

* maybe the last phpstan errors fixed?

* nope, one more fix

* phpstan done, now php-cs-fixer

* fixing two errors found in tests on php 8

* removing lowest version of php-code-coverage, bumping to php-setup v2

* removing other versions of php-code-coverage

* dropping 7.1 from tests, removing php-code-coverage 6.0 that was needed for 7.1

* trying dropping php 7.2 and code-coverage 7.0

* upgrading to php-code-coverage 9.x

* using driver selector for code coverage

* minimum of php-code-coverage 9.1.10 is required for Selector

* removing psr0 php-cs-fixer rule, updating php-cs-fixer

* removing deprecated call to Config::create

* adding back psr0 fixer, upgrading phpstan/php-cs-fixer

the php-cs-fixer team released a fix for the renaming issue I saw with 2.17.2. It still renames tests though, but they're not defined in our autoloader, and the classnames don't match the filename like they should, so excluding those from php-cs-fixer for now.

* allowing symfony yaml 5

* updating psalm/phpstan/php-cs-fixer and ignoring a couple of errors that came up

* updating psalm baseline, new errors did not seem applicable, or easy to fix

+++ lib/Doctrine/Connection/UnitOfWork.php
@@ -54,6 +54,7 @@ class Doctrine_Connection_UnitOfWork extends Doctrine_Connection_Module
         $conn = $this->getConnection();
         $conn->connect();
 
+        /** @var int $state Won't be void since not called with a parameter */
         $state = $record->state();
         if ($state === Doctrine_Record::STATE_LOCKED || $state === Doctrine_Record::STATE_TLOCKED) {
             return false;
@@ -789,7 +790,7 @@ class Doctrine_Connection_UnitOfWork extends Doctrine_Connection_Module
                         }
 
                         unset($flushList[$index3]);
-                        array_splice($flushList, $index - 1, 0, $assocClassName);
+                        array_splice($flushList, (int) ($index - 1), 0, $assocClassName);
                         $index = $relatedCompIndex;
                     } else {
                         $flushList[] = $assocClassName;
